requirements.txt
sounddevice
torch
numpy
coqui-tts
pyaudio